Pretty much all actions now use UTC time.
This fixes a long-standing bug where articles posted in the last 5 hours of the month
would be unviewable.  For example, an article posted at 23:00 Eastern time on Jan 31
would go under February because its posted_at is UTC, but the front page wouldn't link
to February because it used Eastern time to determine the date.