Load articles faster using PJAX.
PJAX lets the client request the article layout without the surrounding
layout, then splice it into the DOM.  This should result in faster page
loads as we don't have to do as many DB accesses.

The PJAX library requires a newer version of jquery, so this patch
upgrades to the latest (1.7.2).