Only use article's charset as a last resort.
It turns out that the charset is less reliable than just guessing.  The
final example was article 25676: an article with content type
"text/plain; charset=ISO-8859-1" with a Swedish signature
containing the word "STÖRSTA" encoded as UTF-8.  Using the
content type's charset just led to wrong results.

It turns out that in the specs we have, the charset is never used.  We
could probably just guess UTF-8 and then ISO-8859-1, and be right 100%
of the time.